+++ src/types-internal/editor-modules.d.ts
@@ -13,7 +13,7 @@ import Tooltip from '../components/modules/tooltip';
 import DragNDrop from '../components/modules/dragNDrop';
 import ModificationsObserver from '../components/modules/modificationsObserver';
 import Renderer from '../components/modules/renderer';
-import Sanitizer from '../components/modules/sanitizer';
+import Sanitizer from '../components/utils/sanitizer';
 import Tools from '../components/modules/tools';
 import API from '../components/modules/api/index';
 import Caret from '../components/modules/caret';